Mobile Item 95: New Competitions Entry Functionality added in MHS GOLF
Slots Listing
- Update UI with 'Add Player', 'Remove (Cross)' or 'already added' according to 'EntryStatus' key.
- New members should be add start from 0th position or after the free location in slots even tap on the last slot.
- Update minor changes in UI.

Members/Friends Selection
- Members/Friends should not be in the listing if already selected.

+++ MHS Demo/MHSystems/app/src/main/java/com/mh/systems/demoapp/web/models/competitionsentrynew/AllPlayer.java
@@ -19,7 +19,7 @@ public class AllPlayer implements Serializable {
     private Boolean IsFriend;
     @SerializedName("EntryFee")
     @Expose
-    private Integer EntryFee;
+    private Double EntryFee;
 
     /**
      * EntryStatus equal to
@@ -57,11 +57,11 @@ public class AllPlayer implements Serializable {
         this.IsFriend = IsFriend;
     }
 
-    public Integer getEntryFee() {
+    public Double getEntryFee() {
         return EntryFee;
     }
 
-    public void setEntryFee(Integer EntryFee) {
+    public void setEntryFee(Double EntryFee) {
         this.EntryFee = EntryFee;
     }
